Game Programming Project Milestone 2
This upload contain the assets and code for Jurgen Famula's Game Programming Project MileStone 2 submittion. I am sorry for turn it in a few hours late, I made some changes to the camera which broke the colliders and pathing for almost everything and I did not notice the problem until after I saved the project.